Compatible json with python < 3.5
As you can see in https://docs.python.org/3/library/json.html#json.JSONDecodeError this exception was added in Python 3.5.

Having to require `simplejson` just for the exception seems a bit too much for the job. It's easier to handle that `ImportError` separately and still allow the native `json` module to work.